Structure for part 2) Process the data.
The structure for processing the data has been coded in the main file. In the Book class, functions have been included but for the moment
have no functionality :generate_book_content, translate_book_content, convert_price_to_euros have "dummy" procedure.

Since the API is not working without paying, for the moment generating the same text "This is a test" for each book.

Translating function has still to be completed.

Also the function converting the price to euros has to be completed.